logicCircuits first version
Mostly this change consists of the class for the "board", it is modelled by a nxm matrix of Gates, at every point we can add or remove gates, there are methods for connecting and unconnecting two gates at a specified input port. We maintain a List of Gates which are not used as input anywhere else, these are our "final output gates". To get the output of the board, we call the output method on these Gates.
------------
In order to be able to unconnect properly, I needed to add a getter for specific input-gates to all the Gate-classes